requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
